WIP
I did way to much here without committing, and I'm not sure any specific
mutation is complete. Going to pick it apart and recreate it one step (and
commit) at a time on master, but I wanted it in the repo on a branch for both
history and to be able to transfer it between computers.